Rename main directory to mainappsrc
